The Code for the detection.
The code involves all important libraries along with the Haar cascade classifiers which is crucial for smile detection. The first step is to detect the dace followed by the smile detection. The final step will be to calculate the ratio of the width of the smile in the image to the width of the face itself. This ratio is later converted to a categorical value between 0 and 100.